Moved code to src/ and added docker info.
This is mostly a code reorganization to simplify building the
docker image.

It also has a small fix for xss checking in src/Library/Bootstrap.php
that recursively checks values that are passed in as an array instead
of skipping them.

This is a first pass to make it easy for developers to get this up
and running on their local machines without needing to know too
much about how to set up a php application.